requirements/prod.txt
-r ../vendor/src/funfactory/funfactory/requirements/prod.txt

djangorestframework==2.3.10
django_ace==1.0.0
django-braces==1.2.2
django-filter==0.7
django-mptt==0.6.0
markdown==2.3.1
south==0.8.2

# Requirements tied to specific commits
git+git://github.com/jsocol/jingo-minify.git@2beb546949ed2cb4eb866bab6dd643047d803a95
git+git://github.com/mozilla/django-browserid.git@5a0eee5c3e315daf603c821a90ccc5fa7d43a0c5
